requirements.txt
Django==1.11
djangorestframework==3.6.2
django_jsonfield==1.0.1
celery==4.0.2
redis==2.10.5
requests==2.13.0
responses==0.5.1
publicsuffix=1.1.0
